fix: move unstable declarations to deno.unstable (#11876)
