+++ cli/main.rs
@@ -358,8 +358,6 @@ pub fn get_types(unstable: bool) -> String {
 
   if unstable {
     types.push(crate::tsc::UNSTABLE_NS_LIB);
-    types.push(crate::tsc::DENO_NET_UNSTABLE_LIB);
-    types.push(crate::tsc::DENO_HTTP_UNSTABLE_LIB);
   }
 
   types.join("\n")
